+++ backend/src/routes/index.ts
@@ -3,10 +3,8 @@ import {
   createTodo,
   getAllTodos,
   getTodoById,
-  updateTodoById,
-  deleteTodoById,
-  createRandomTodo,
-} from "./Todos";
+} from "./methods";
+import { deleteTodoById, updateTodoById } from "./newMethods";
 
 const todoRouter = Router();
 todoRouter.post("/todos", createTodo);
@@ -14,6 +12,5 @@ todoRouter.get("/todos", getAllTodos);
 todoRouter.get("/todos/:id", getTodoById);
 todoRouter.put("/todos/:id", updateTodoById);
 todoRouter.delete("/todos/:id", deleteTodoById);
-todoRouter.post("/todos/random", createRandomTodo);
 
 export default todoRouter;
